Move `...Model` types to the corresponding `...Page` files.
